refactor : bookTab 폴더 위치 변경
#636
